requirements.txt
asgiref==3.5.2
boto3==1.26.64
botocore==1.29.64
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
defusedxml==0.7.1
Django==3.2.2
django-allauth==0.51.0
django-cors-headers==3.13.0
django-rest-auth==0.9.5
django-storages==1.13.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
idna==3.3
jmespath==1.0.1
oauthlib==3.2.0
Pillow==9.1.1
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
requests==2.28.0
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
